+++ okhttp/src/main/java/okhttp3/OkHttpClient.kt
@@ -34,6 +34,7 @@ import okhttp3.Protocol.HTTP_2
 import okhttp3.internal.asFactory
 import okhttp3.internal.checkDuration
 import okhttp3.internal.concurrent.TaskRunner
+import okhttp3.internal.connection.RealCall
 import okhttp3.internal.connection.RouteDatabase
 import okhttp3.internal.immutableListOf
 import okhttp3.internal.platform.Platform
@@ -242,9 +243,7 @@ open class OkHttpClient internal constructor(
   }
 
   /** Prepares the [request] to be executed at some point in the future. */
-  override fun newCall(request: Request): Call {
-    return RealCall.newRealCall(this, request, forWebSocket = false)
-  }
+  override fun newCall(request: Request): Call = RealCall(this, request, forWebSocket = false)
 
   /** Uses [request] to connect a new web socket. */
   override fun newWebSocket(request: Request, listener: WebSocketListener): WebSocket {
